fix(store, dashboard): update store to initialize goals object and currentProgress array in state
Update store with actions to initialize state upon new user registration, lifted up logic from
components to be performed during store actions, removed some console.logs.

'Daily Details Error due to Null dailyTotal in store'
