+++ screens/DailyDetails/DailyDetails.jsx
@@ -8,8 +8,6 @@ import { GoalBar } from '../../components/GoalBar/GoalBar';
 export default function DailyDetails() {
   const { dailyTotal, totalGoalMet } = useSelector(state => state);
 
-  console.log('Here in the daily details screen', dailyTotal);
-
   const dateOptions = {
     weekday: 'long',
     month: 'long',
