+++ screens/Dashboard/Dashboard.jsx
@@ -11,40 +11,7 @@ const fact = 'There is no perfect diet for everyone.';
 
 export default function Dashboard({ navigation }) {
 
-  const [metGoal, setMetGoal] = useState(0);
-  const dispatch = useDispatch();
-
-  const {
-    user: {
-      sex,
-      birthdate,
-    },
-    currentProgress,
-  } = useSelector(state => state);
-
-  useEffect(() => {
-    setMetGoal(calcProgress(currentProgress));
-  }, []);
-
-  const calcProgress = (items) => {
-    if (items.length) {
-      const dailyTotal = calcTotalsByNutrient({
-        items,
-        sex,
-        birthdate,
-      });
-      const totalGoalMet = calcTotalProgress(dailyTotal);
-      dispatch({
-        type: ActionType.ADD_DAILY_TOTAL,
-        payload: {
-          dailyTotal,
-          totalGoalMet,
-        }
-      });
-      return totalGoalMet;
-    }
-    return 0;
-  }
+  const { totalGoalMet } = useSelector(state => state);
 
   return (
     <View style={ style.container }>
@@ -52,7 +19,7 @@ export default function Dashboard({ navigation }) {
         <Text style={ style.buttontext }>{ fact} </Text>
       </View>
       <View style={ style.goalBubble }>
-        <Text style={ style.percentage }>{ metGoal }%</Text>
+        <Text style={ style.percentage }>{ totalGoalMet }%</Text>
         <Text style={ style.bubbleText }>of your daily needs have been met!</Text>
       </View>
       <TouchableOpacity style={ style.submitbutton } onPress={ () => navigation.push('Details') }>
